Update free-courses-en.md
This Kotlin tutorial is intended for complete beginners.
You will learn basic programming concepts in Kotlin. Start watching this course if you are interested in learning Kotlin as a beginner.
In this Kotlin tutorial, you will go through some of the basic concepts of the language. Feel free to start watching this tutorial. As a beginner in Kotlin, you will definitely learn some concepts that will enable you to step onto the next concepts. The course is totally free.